Introduce AbstractAgentSet class and refactor imports for consistency
