+++ mesa_frames/concrete/model.py
@@ -46,7 +46,7 @@ from collections.abc import Sequence
 
 import numpy as np
 
-from mesa_frames.abstract.agentsetregistry import AbstractAgentSet
+from mesa_frames.abstract.agentset import AbstractAgentSet
 from mesa_frames.abstract.space import SpaceDF
 from mesa_frames.concrete.agentsetregistry import AgentSetRegistry
